requirements.txt
pygame>=1.9.2b6
six>=1.10.0
websocket-client>=0.37.0
websockets>=3.2
